data_table elements moved to separate
