+++ dashboards-observability/common/constants/data_table.ts
@@ -4,7 +4,7 @@
  */
 
 export const GRID_HEADER_COLUMN_MAX_WIDTH = '150px';
-export const GRID_PAGE_RANGE_DSIPLAY = 5;
+export const GRID_PAGE_RANGE_DISPLAY = 5;
 export const COLUMN_DEFAULT_MIN_WIDTH = 100;
 export const GRID_PAGE_SIZES = [10, 50, 100];
 export const ROW_DENSITIES = [
