+++ dashboards-observability/opensearch_dashboards.json
@@ -1,7 +1,7 @@
 {
   "id": "observabilityDashboards",
-  "version": "2.1.0.0",
-  "opensearchDashboardsVersion": "2.1.0",
+  "version": "2.0.0.0",
+  "opensearchDashboardsVersion": "2.0.0",
   "server": true,
   "ui": true,
   "requiredPlugins": [
