+++ dashboards-observability/package.json
@@ -1,6 +1,6 @@
 {
   "name": "observability-dashboards",
-  "version": "2.1.0.0",
+  "version": "2.0.0.0",
   "main": "index.ts",
   "license": "Apache-2.0",
   "scripts": {
